requirements.txt
folium==0.17.0
matplotlib-inline==0.1.7
pandas==2.0.3
plotly==5.24.0
streamlit==1.38.0
streamlit_folium==0.22.0
snowflake-connector-python==3.12.2
snowflake-ml-python==1.6.2
snowflake-snowpark-python==1.22.1
streamlit==1.38.0
streamlit-folium==0.22.0
streamlit-pdf-viewer==0.0.18
matplotlib==3.7.5
python-dotenv==1.0.1



